Import upstream VSCode 1.129.0 (replace tree with microsoft/vscode tag 1.129.0)

+++ extensions/typescript-language-features/src/tsServer/serverProcess.electron.ts
@@ -11,11 +11,11 @@ import * as vscode from 'vscode';
 import { TypeScriptServiceConfiguration } from '../configuration/configuration';
 import { Disposable } from '../utils/dispose';
 import { API } from './api';
-import { NodeVersionManager } from './nodeManager';
 import type * as Proto from './protocol/protocol';
 import { TsServerLog, TsServerProcess, TsServerProcessFactory, TsServerProcessKind } from './server';
 import { TypeScriptVersionManager } from './versionManager';
 import { TypeScriptVersion } from './versionProvider';
+import { NodeVersionManager } from './nodeManager';
 
 
 const defaultSize: number = 8192;
@@ -24,6 +24,12 @@ const contentLengthSize: number = Buffer.byteLength(contentLength, 'utf8');
 const blank: number = Buffer.from(' ', 'utf8')[0];
 const backslashR: number = Buffer.from('\r', 'utf8')[0];
 const backslashN: number = Buffer.from('\n', 'utf8')[0];
+const gracefulExitTimeout = 5000;
+const tsServerExitRequest: Proto.Request = {
+	seq: 0,
+	type: 'request',
+	command: 'exit',
+};
 
 class ProtocolBuffer {
 
@@ -38,14 +44,14 @@ class ProtocolBuffer {
 			toAppend = Buffer.from(data, 'utf8');
 		}
 		if (this.buffer.length - this.index >= toAppend.length) {
-			toAppend.copy(this.buffer as unknown as Uint8Array, this.index, 0, toAppend.length);
+			toAppend.copy(this.buffer, this.index, 0, toAppend.length);
 		} else {
 			const newSize = (Math.ceil((this.index + toAppend.length) / defaultSize) + 1) * defaultSize;
 			if (this.index === 0) {
 				this.buffer = Buffer.allocUnsafe(newSize);
-				toAppend.copy(this.buffer as unknown as Uint8Array, 0, 0, toAppend.length);
+				toAppend.copy(this.buffer, 0, 0, toAppend.length);
 			} else {
-				this.buffer = Buffer.concat([this.buffer.slice(0, this.index), toAppend] as Uint8Array[], newSize);
+				this.buffer = Buffer.concat([this.buffer.slice(0, this.index), toAppend], newSize);
 			}
 		}
 		this.index += toAppend.length;
@@ -85,7 +91,7 @@ class ProtocolBuffer {
 		while (sourceStart < this.index && (this.buffer[sourceStart] === backslashR || this.buffer[sourceStart] === backslashN)) {
 			sourceStart++;
 		}
-		this.buffer.copy(this.buffer as unknown as Uint8Array, 0, sourceStart);
+		this.buffer.copy(this.buffer, 0, sourceStart);
 		this.index = this.index - sourceStart;
 		return result;
 	}
@@ -162,6 +168,24 @@ function getExecArgv(kind: TsServerProcessKind, configuration: TypeScriptService
 		args.push(`--max-old-space-size=${configuration.maxTsServerMemory}`);
 	}
 
+	if (configuration.diagnosticDir) {
+		args.push(`--diagnostic-dir=${configuration.diagnosticDir}`);
+	}
+
+	if (configuration.heapSnapshot > 0) {
+		args.push(`--heapsnapshot-near-heap-limit=${configuration.heapSnapshot}`);
+	}
+
+	if (configuration.heapProfile.enabled) {
+		args.push('--heap-prof');
+		if (configuration.heapProfile.dir) {
+			args.push(`--heap-prof-dir=${configuration.heapProfile.dir}`);
+		}
+		if (configuration.heapProfile.interval) {
+			args.push(`--heap-prof-interval=${configuration.heapProfile.interval}`);
+		}
+	}
+
 	return args;
 }
 
@@ -189,10 +213,15 @@ function getTssDebugBrk(): string | undefined {
 }
 
 class IpcChildServerProcess extends Disposable implements TsServerProcess {
+	private _killTimeout: NodeJS.Timeout | undefined;
+	private _isShuttingDown = false;
+
 	constructor(
 		private readonly _process: child_process.ChildProcess,
+		private readonly _useGracefulShutdown: boolean,
 	) {
 		super();
+		this._process.once('exit', () => this.clearKillTimeout());
 	}
 
 	write(serverRequest: Proto.Request): void {
@@ -212,18 +241,47 @@ class IpcChildServerProcess extends Disposable implements TsServerProcess {
 	}
 
 	kill(): void {
-		this._process.kill();
+		if (!this._useGracefulShutdown) {
+			this._process.kill();
+			return;
+		}
+
+		if (this._isShuttingDown) {
+			return;
+		}
+		this._isShuttingDown = true;
+
+		try {
+			this._process.send(tsServerExitRequest);
+		} catch {
+			this._process.kill();
+			return;
+		}
+
+		this._killTimeout = setTimeout(() => this._process.kill(), gracefulExitTimeout);
+		this._killTimeout.unref?.();
+	}
+
+	private clearKillTimeout(): void {
+		if (this._killTimeout) {
+			clearTimeout(this._killTimeout);
+			this._killTimeout = undefined;
+		}
 	}
 }
 
 class StdioChildServerProcess extends Disposable implements TsServerProcess {
 	private readonly _reader: Reader<Proto.Response>;
+	private _killTimeout: NodeJS.Timeout | undefined;
+	private _isShuttingDown = false;
 
 	constructor(
 		private readonly _process: child_process.ChildProcess,
+		private readonly _useGracefulShutdown: boolean,
 	) {
 		super();
 		this._reader = this._register(new Reader<Proto.Response>(this._process.stdout!));
+		this._process.once('exit', () => this.clearKillTimeout());
 	}
 
 	write(serverRequest: Proto.Request): void {
@@ -244,7 +302,39 @@ class StdioChildServerProcess extends Disposable implements TsServerProcess {
 	}
 
 	kill(): void {
-		this._process.kill();
+		if (!this._useGracefulShutdown) {
+			this._process.kill();
+			this._reader.dispose();
+			return;
+		}
+
+		if (this._isShuttingDown) {
+			return;
+		}
+		this._isShuttingDown = true;
+
+		try {
+			this._process.stdin?.write(JSON.stringify(tsServerExitRequest) + '\r\n', 'utf8');
+			this._process.stdin?.end();
+		} catch {
+			this._process.kill();
+			this._reader.dispose();
+			return;
+		}
+
+		this._killTimeout = setTimeout(() => {
+			this._process.kill();
+			this._reader.dispose();
+		}, gracefulExitTimeout);
+		this._killTimeout.unref?.();
+	}
+
+	private clearKillTimeout(): void {
+		if (this._killTimeout) {
+			clearTimeout(this._killTimeout);
+			this._killTimeout = undefined;
+		}
+
 		this._reader.dispose();
 	}
 }
@@ -272,6 +362,7 @@ export class ElectronServiceProcessFactory implements TsServerProcessFactory {
 		const env = generatePatchedEnv(process.env, tsServerPath, !!execPath);
 		const runtimeArgs = [...args];
 		const execArgv = getExecArgv(kind, configuration);
+		const useGracefulShutdown = configuration.heapProfile.enabled;
 		const useIpc = !execPath && version.apiVersion?.gte(API.v460);
 		if (useIpc) {
 			runtimeArgs.push('--useNodeIpc');
@@ -291,6 +382,6 @@ export class ElectronServiceProcessFactory implements TsServerProcessFactory {
 				stdio: useIpc ? ['pipe', 'pipe', 'pipe', 'ipc'] : undefined,
 			});
 
-		return useIpc ? new IpcChildServerProcess(childProcess) : new StdioChildServerProcess(childProcess);
+		return useIpc ? new IpcChildServerProcess(childProcess, useGracefulShutdown) : new StdioChildServerProcess(childProcess, useGracefulShutdown);
 	}
 }
